canonicalize import order
fixes smlnj/smlnj#319

The import tuple is sorted by the PIDs. This change canonicalizes the
order to be based on the appearance order of the external value in the
source code. Specifically, names from the same external CU is still
grouped together, but when the first value from a CU appears, the CU is
given a number that is used in sorting.