fix unification errors
fixes #317

The previous implementation does not retry unification after a failure,
but sometimes the types are unifiable after a head-reduction. This
commit ports the retrying code from the legacy branch. In addition, when
instantiating/unifying two overload tyvars, the direction of unification
matters for overload resolution, which the previous implementation did
not take into account.